requirements.txt
Flask
pdfplumber
gTTSg
gunicorn